Gb/feature/matriz

+++ src/main/java/school/sptech/backend/service/condominio/CondominioService.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 import school.sptech.backend.domain.condominio.Condominio;
 import school.sptech.backend.domain.condominio.repository.CondominioRepository;
+import school.sptech.backend.exception.NaoEncontradoException;
+import school.sptech.backend.service.endereco.EnderecoService;
 
 import java.util.List;
 
@@ -15,45 +17,42 @@ public class CondominioService {
 
     private final CondominioRepository repository;
 
-    public Condominio criar(Condominio condominio) {
+    private final EnderecoService enderecoService;
+
+    public Condominio criar(Condominio condominio, Integer enderecoId) {
+        condominio.setEndereco(enderecoService.porId(enderecoId));
         return this.repository.save(condominio);
     }
 
     public List<Condominio> listar() {
-        final List<Condominio> condominios = this.repository.findAll();
-
-        if (condominios.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NO_CONTENT);
-        }
-
-        return condominios;
+        return this.repository.findAll();
     }
 
     public Condominio porId(int id) {
         return this.repository.findById(id).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
+                () -> new NaoEncontradoException("Condomínio")
         );
     }
 
     public Condominio porNome(String nome) {
         return this.repository.findByNomeIgnoreCase(nome).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
+                () -> new NaoEncontradoException("Condomínio")
         );
     }
 
     public Condominio atualizar(Condominio condominioAtualizado, int id) {
-        this.repository.findById(id).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
-        );
+        if (!this.repository.existsById(id)) {
+            throw new NaoEncontradoException("Condomínio");
+        }
 
         condominioAtualizado.setId(id);
         return this.repository.save(condominioAtualizado);
     }
 
     public void deletar(int id) {
-        this.repository.findById(id).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
-        );
+        if (!this.repository.existsById(id)) {
+            throw new NaoEncontradoException("Condomínio");
+        }
 
         this.repository.deleteById(id);
     }
